support pc development with JS http bridge

+++ pc/device.c
@@ -95,12 +95,12 @@ void udp_close(int fd)
 
 
 
-uint64_t millis()
+uint32_t millis()
 {
     struct timeval te;
     gettimeofday(&te, NULL); // get current time
     uint64_t milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // calculate milliseconds
-    return milliseconds;
+    return (uint32_t)milliseconds;
 }
